:truck: move thanks into autprespond
