+++ src/v2/autorespond/thanks/checker.test.ts
